Draw.io File for Electrical Setup
So in this diagram I am including the control system, battery charging based on MPPT, along with a couple of electronic switches I placed in a few different spots.
The 1st electronic switch is connected between the solar array and the MPPT charging circuit. This will allow us to shutoff the array in case of plane failure or if we need to halt charging or what not.