Current Shunt Basic Idea Schematic
This is an LTSpice schematic for a rough idea I have going for a Current shunt using a high-power resistor and a logic level mosfet controlled by 5V logic to drain excess current from our system. The goal here would be that based on our battery max current and measurements of current draw and gain we can tell if we are having too high of current charging the battery which we can decrease by increasing the voltage at the gate of the mosfet. I know this sort of wastes some of the current but I believe this would leave it modular so if we did have a better battery capacity we just change the values in our controller and basically use the resistor more or less depending on current requirements.